requirements.txt

pyyaml
funcy
imagesize
ffmpeg-python
fleep
# instagram_private_api @ git+https://github.com/ping/instagram_private_api.git@1.6.0
colorama
colorlog
click
modeller
portalocker